+++ packages/sdk/src/sdk-blockchains/tezos/cancel.test.ts
@@ -1,32 +1,24 @@
 // eslint-disable-next-line camelcase
-import { toBigNumber, toCollectionId, toContractAddress } from "@rarible/types"
+import { toCollectionId } from "@rarible/types"
 import { createRaribleSdk } from "../../index"
 import { MintType } from "../../types/nft/mint/domain"
-import { delay, retry } from "../../common/retry"
+import { retry } from "../../common/retry"
 import { LogsLevel } from "../../domain"
-import { createApisSdk } from "../../common/apis"
-import { awaitForItemSupply } from "./test/await-for-item-supply"
+import type { RaribleSdkEnvironment } from "../../config/domain"
+import { awaitItemSupply } from "../ethereum/test/await-item-supply"
 import { createTestWallet } from "./test/test-wallet"
-import { getMaybeTezosProvider, getTezosAPIs } from "./common"
-import { TezosSell } from "./sell"
+import { getTestContract } from "./test/test-contracts"
 
-describe.skip("cancel test", () => {
-	const wallet = createTestWallet("edsk3UUamwmemNBJgDvS8jXCgKsvjL2NoTwYRFpGSRPut4Hmfs6dG8")
-	const env = "development"
+describe("cancel test", () => {
+	const env: RaribleSdkEnvironment = "testnet"
+	const wallet = createTestWallet("edsk3UUamwmemNBJgDvS8jXCgKsvjL2NoTwYRFpGSRPut4Hmfs6dG8", env)
 	const sdk = createRaribleSdk(wallet, env, { logs: LogsLevel.DISABLED })
 
-	const tezosNetwork = "dev"
-	const sellerTezosProvider = getMaybeTezosProvider(wallet.provider, tezosNetwork)
-	const apis = getTezosAPIs(tezosNetwork)
-	const unionApis = createApisSdk(env, undefined)
-	const sellerSellService = new TezosSell(sellerTezosProvider, apis, unionApis)
-
-	let nftContract: string = "KT1PuABq2ReD789KtKetktvVKJcCMpyDgwUx"
-	const eurTzContract = "KT1HvTfYG7DgeujAQ1LDvCHiQc29VMycoJh5"
+	const nftContract: string = getTestContract(env, "nftContract")
 
 	test("cancel order", async () => {
 		const mintResponse = await sdk.nft.mint({
-			collectionId: toCollectionId(`TEZOS:${nftContract}`),
+			collectionId: toCollectionId(nftContract),
 		})
 		const mintResult = await mintResponse.submit({
 			uri: "ipfs://bafkreiaz7n5zj2qvtwmqnahz7rwt5h37ywqu7znruiyhwuav3rbbxzert4",
@@ -37,7 +29,7 @@ describe.skip("cancel test", () => {
 			await mintResult.transaction.wait()
 		}
 
-		await awaitForItemSupply(sdk, mintResult.itemId, "1")
+		await awaitItemSupply(sdk, mintResult.itemId, "1")
 
 		const sellAction = await sdk.order.sell({
 			itemId: mintResult.itemId,
@@ -51,7 +43,7 @@ describe.skip("cancel test", () => {
 			},
 		})
 
-		await retry(10, 1000, async () => {
+		await retry(20, 2000, async () => {
 			const order = await sdk.apis.order.getOrderById({
 				id: orderId,
 			})
@@ -60,66 +52,12 @@ describe.skip("cancel test", () => {
 			}
 		})
 
-		await delay(10000)
+		// await delay(5000)
 		const cancelTx = await sdk.order.cancel({
 			orderId,
 		})
 		await cancelTx.wait()
-		await retry(10, 1000, async () => {
-			const canceledOrder = await sdk.apis.order.getOrderById({
-				id: orderId,
-			})
-			if (canceledOrder.status !== "CANCELLED") {
-				throw new Error("Order has not been cancelled")
-			}
-		})
-
-
-	}, 1500000)
-
-	test("cancel v1 order", async () => {
-		const mintResponse = await sdk.nft.mint({
-			collectionId: toCollectionId(`TEZOS:${nftContract}`),
-		})
-		const mintResult = await mintResponse.submit({
-			uri: "ipfs://bafkreiaz7n5zj2qvtwmqnahz7rwt5h37ywqu7znruiyhwuav3rbbxzert4",
-			supply: 1,
-			lazyMint: false,
-		})
-		if (mintResult.type === MintType.ON_CHAIN) {
-			await mintResult.transaction.wait()
-		}
-
-		await awaitForItemSupply(sdk, mintResult.itemId, "1")
-
-		const orderId = await sellerSellService.sellV1({
-			itemId: mintResult.itemId,
-			amount: 1,
-			price: "0.002",
-			currency: {
-				"@type": "TEZOS_FT",
-				contract: toContractAddress(
-					`TEZOS:${eurTzContract}`
-				),
-				tokenId: toBigNumber("0"),
-			},
-		})
-
-		await retry(10, 1000, async () => {
-			const order = await sdk.apis.order.getOrderById({
-				id: orderId,
-			})
-			if (order.status !== "ACTIVE") {
-				throw new Error("Order status is not active")
-			}
-		})
-
-		await delay(10000)
-		const cancelTx = await sdk.order.cancel({
-			orderId,
-		})
-		await cancelTx.wait()
-		await retry(10, 1000, async () => {
+		await retry(40, 2000, async () => {
 			const canceledOrder = await sdk.apis.order.getOrderById({
 				id: orderId,
 			})