+++ packages/sdk/src/sdk-blockchains/tezos/create-collection.test.ts
@@ -1,15 +1,18 @@
 import { Blockchain } from "@rarible/api-client"
 import { createRaribleSdk } from "../../index"
 import { LogsLevel } from "../../domain"
+import type { RaribleSdkEnvironment } from "../../config/domain"
 import { awaitForCollection } from "./test/await-for-collection"
 import { createTestWallet } from "./test/test-wallet"
 
-describe.skip("deploy tezos tests", () => {
+describe("deploy tezos tests", () => {
+	const env: RaribleSdkEnvironment = "testnet"
 	const wallet = createTestWallet(
 		"edskRqrEPcFetuV7xDMMFXHLMPbsTawXZjH9yrEz4RBqH1" +
-    "D6H8CeZTTtjGA3ynjTqD8Sgmksi7p5g3u5KUEVqX2EWrRnq5Bymj"
+    "D6H8CeZTTtjGA3ynjTqD8Sgmksi7p5g3u5KUEVqX2EWrRnq5Bymj",
+		env
 	)
-	const sdk = createRaribleSdk(wallet, "development", { logs: LogsLevel.DISABLED })
+	const sdk = createRaribleSdk(wallet, env, { logs: LogsLevel.DISABLED })
 
 	test("deploy public nft", async () => {
 		const result = await sdk.nft.createCollection({
@@ -62,7 +65,6 @@ describe.skip("deploy tezos tests", () => {
 		})
 
 		await result.tx.wait()
-		console.log("result", result.address)
 		await awaitForCollection(sdk, result.address)
 	})