+++ CHANGELOG.md
@@ -3,6 +3,132 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.11-beta.13](https://github.com/rarible/sdk/compare/v0.9.11-beta.12...v0.9.11-beta.13) (2022-07-08)
+
+
+### Features
+
+* seaport orders ([#287](https://github.com/rarible/sdk/issues/287)) ([6fc0e9c](https://github.com/rarible/sdk/commit/6fc0e9caa030ba835abe6f9eb9ccc3827f78ea3f))
+
+
+
+
+
+## [0.9.11-beta.12](https://github.com/rarible/sdk/compare/v0.9.11-beta.11...v0.9.11-beta.12) (2022-07-06)
+
+
+### Bug Fixes
+
+* tezos buy price for multiple items order ([225df7c](https://github.com/rarible/sdk/commit/225df7ccd10a20f86c7d3ac454c466ea77e9cd32))
+* tezos buy price for multiple items order ([e8779d6](https://github.com/rarible/sdk/commit/e8779d6f87f88a2707882ac80fd79c76d066debb))
+
+
+
+
+
+## [0.9.11-beta.11](https://github.com/rarible/sdk/compare/v0.9.11-beta.10...v0.9.11-beta.11) (2022-07-04)
+
+
+### Bug Fixes
+
+* correct use baseFee info from auction house's mapping ([8b66aea](https://github.com/rarible/sdk/commit/8b66aea7e360fa85487db57742791cb26b6b5643))
+
+
+### Features
+
+* extend e2e tests coverage ([0f77508](https://github.com/rarible/sdk/commit/0f775087a2b00d620e517c82eeadd70ba0445e7c))
+* extend e2e tests coverage ([27dcd3c](https://github.com/rarible/sdk/commit/27dcd3c72451b19d9ae60c389262ccff8fa5c8e3))
+
+
+
+
+
+## [0.9.11-beta.10](https://github.com/rarible/sdk/compare/v0.9.11-beta.9...v0.9.11-beta.10) (2022-06-30)
+
+
+### Bug Fixes
+
+* tezos changes ([76cc534](https://github.com/rarible/sdk/commit/76cc534ac6b0f3232c5f242feaddc10dc43190cd))
+
+
+
+
+
+## [0.9.11-beta.9](https://github.com/rarible/sdk/compare/v0.9.11-beta.8...v0.9.11-beta.9) (2022-06-29)
+
+
+### Bug Fixes
+
+* opensea order payouts ([002e901](https://github.com/rarible/sdk/commit/002e90161f5586d1ab9572fccc091b112ca5e12d))
+
+
+
+
+
+## [0.9.11-beta.8](https://github.com/rarible/sdk/compare/v0.9.11-beta.7...v0.9.11-beta.8) (2022-06-27)
+
+
+### Bug Fixes
+
+* tezos sell v2 on dev ([b15235f](https://github.com/rarible/sdk/commit/b15235fc052cdf94eaff7b0a02f0e311ca561b67))
+
+
+
+
+
+## [0.9.11-beta.7](https://github.com/rarible/sdk/compare/v0.9.11-beta.6...v0.9.11-beta.7) (2022-06-23)
+
+
+### Bug Fixes
+
+* tezos prod order versions ([c998266](https://github.com/rarible/sdk/commit/c998266b340e5200877a0f116de3f091ff6466e2))
+
+
+
+
+
+## [0.9.11-beta.6](https://github.com/rarible/sdk/compare/v0.9.11-beta.5...v0.9.11-beta.6) (2022-06-22)
+
+
+### Bug Fixes
+
+* dipdup for tezos testnet ([aefa1f4](https://github.com/rarible/sdk/commit/aefa1f4c2f830555fa1bc006858cb5e5491c1ffc))
+
+
+
+
+
+## [0.9.11-beta.5](https://github.com/rarible/sdk/compare/v0.9.11-beta.4...v0.9.11-beta.5) (2022-06-22)
+
+
+### Bug Fixes
+
+* union api url for testnet ([#267](https://github.com/rarible/sdk/issues/267)) ([1ee426a](https://github.com/rarible/sdk/commit/1ee426a3417c2f38e843f2b55caf22a00a3fd61f))
+
+
+
+
+
+## [0.9.11-beta.4](https://github.com/rarible/sdk/compare/v0.9.11-beta.2...v0.9.11-beta.4) (2022-06-22)
+
+
+### Bug Fixes
+
+* add retry policy for solana tx confirmation ([3d25ab0](https://github.com/rarible/sdk/commit/3d25ab07e57e7b34464cab1bb4f349925e7aef08))
+* set longer default timeout for solana txs ([48b4e0f](https://github.com/rarible/sdk/commit/48b4e0f020d9d2627248346f956336b39db3ee72))
+* solana burn & transfer ([2f894d8](https://github.com/rarible/sdk/commit/2f894d8a71efffa167d9eef222aa5ba4f81a07e2))
+* solana burn & transfer ([dfdaee8](https://github.com/rarible/sdk/commit/dfdaee869f173f554ab840e5e654ca27d71018df))
+* solana cancel for multiple items ([e55fa39](https://github.com/rarible/sdk/commit/e55fa394ffdd78f88990f92e2409b3083a6d733e))
+
+
+### Features
+
+* add testnet env, remove e2e env ([#261](https://github.com/rarible/sdk/issues/261)) ([55c908f](https://github.com/rarible/sdk/commit/55c908f031e39b8a2ad504c1e334695d3386ce42))
+
+
+
+
+
 ## [0.9.11-beta.2](https://github.com/rarible/sdk/compare/v0.9.11-beta.1...v0.9.11-beta.2) (2022-06-20)