+++ packages/sdk-examples/CHANGELOG.md
@@ -3,6 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.11-beta.13](https://github.com/rarible/sdk/compare/v0.9.11-beta.12...v0.9.11-beta.13) (2022-07-08)
+
+
+### Features
+
+* seaport orders ([#287](https://github.com/rarible/sdk/issues/287)) ([6fc0e9c](https://github.com/rarible/sdk/commit/6fc0e9caa030ba835abe6f9eb9ccc3827f78ea3f))
+
+
+
+
+
+## [0.9.11-beta.12](https://github.com/rarible/sdk/compare/v0.9.11-beta.11...v0.9.11-beta.12) (2022-07-06)
+
+
+### Bug Fixes
+
+* tezos buy price for multiple items order ([225df7c](https://github.com/rarible/sdk/commit/225df7ccd10a20f86c7d3ac454c466ea77e9cd32))
+
+
+
+
+
+## [0.9.11-beta.11](https://github.com/rarible/sdk/compare/v0.9.11-beta.10...v0.9.11-beta.11) (2022-07-04)
+
+**Note:** Version bump only for package @rarible/sdk-examples
+
+
+
+
+
+## [0.9.11-beta.10](https://github.com/rarible/sdk/compare/v0.9.11-beta.9...v0.9.11-beta.10) (2022-06-30)
+
+
+### Bug Fixes
+
+* tezos changes ([76cc534](https://github.com/rarible/sdk/commit/76cc534ac6b0f3232c5f242feaddc10dc43190cd))
+
+
+
+
+
+## [0.9.11-beta.9](https://github.com/rarible/sdk/compare/v0.9.11-beta.8...v0.9.11-beta.9) (2022-06-29)
+
+**Note:** Version bump only for package @rarible/sdk-examples
+
+
+
+
+
+## [0.9.11-beta.8](https://github.com/rarible/sdk/compare/v0.9.11-beta.7...v0.9.11-beta.8) (2022-06-27)
+
+**Note:** Version bump only for package @rarible/sdk-examples
+
+
+
+
+
+## [0.9.11-beta.7](https://github.com/rarible/sdk/compare/v0.9.11-beta.6...v0.9.11-beta.7) (2022-06-23)
+
+
+### Bug Fixes
+
+* tezos prod order versions ([c998266](https://github.com/rarible/sdk/commit/c998266b340e5200877a0f116de3f091ff6466e2))
+
+
+
+
+
+## [0.9.11-beta.6](https://github.com/rarible/sdk/compare/v0.9.11-beta.5...v0.9.11-beta.6) (2022-06-22)
+
+**Note:** Version bump only for package @rarible/sdk-examples
+
+
+
+
+
+## [0.9.11-beta.5](https://github.com/rarible/sdk/compare/v0.9.11-beta.4...v0.9.11-beta.5) (2022-06-22)
+
+**Note:** Version bump only for package @rarible/sdk-examples
+
+
+
+
+
+## [0.9.11-beta.4](https://github.com/rarible/sdk/compare/v0.9.11-beta.2...v0.9.11-beta.4) (2022-06-22)
+
+
+### Features
+
+* add testnet env, remove e2e env ([#261](https://github.com/rarible/sdk/issues/261)) ([55c908f](https://github.com/rarible/sdk/commit/55c908f031e39b8a2ad504c1e334695d3386ce42))
+
+
+
+
+
 ## [0.9.11-beta.2](https://github.com/rarible/sdk/compare/v0.9.11-beta.1...v0.9.11-beta.2) (2022-06-20)
 
 **Note:** Version bump only for package @rarible/sdk-examples